+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The rules below are promoted from the Experiment ruleset to the Standard ruleset
 * `package-name`
 * `traling-comma`
 * `spacing-around-angle-brackets`
+* `unary-op-spacing`
 
 Note that as a result of moving the rules that the prefix `experimental:` has to be removed from all references to this rule. Check references in:
 * The `.editorconfig` setting `disabled_rules`.
