+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/SpacingAroundUnaryOperatorRuleTest.kt
@@ -1,6 +1,5 @@
-package com.pinterest.ktlint.ruleset.experimental
+package com.pinterest.ktlint.ruleset.standard
 
-import com.pinterest.ktlint.ruleset.standard.SpacingAroundUnaryOperatorRule
 import com.pinterest.ktlint.test.KtLintAssertThat.Companion.assertThat
 import com.pinterest.ktlint.test.LintViolation
 import org.assertj.core.api.Assertions.assertThat
